feat: added basket list

+++ Dto/BasketDto/ResultBasketDto.cs
@@ -9,9 +9,8 @@ namespace Dto.BasketDto
     public class ResultBasketDto
     {
         public decimal BasketID { get; set; }
-
         public decimal Price { get; set; }
-        public decimal Count { get; set; }
+        public int Count { get; set; }
         public decimal TotalPrice { get; set; }
         public int ProductID { get; set; }
         public int MenuTableID { get; set; }
